+++ lib/utils/terminal_colors.dart
@@ -2,4 +2,5 @@ const red = '\x1B[91m';
 const green = '\x1B[92m';
 const yellow = '\x1B[93m';
 const orange = '\x1B[38;5;208m';
+const cyan = '\x1B[96m';
 const reset = '\x1B[0m';